Essentially the code was set up so that the user can search for songs and queue them through our APP. After doing research and discussing it amongst the group, we've decided that it would be not only easier, but helpful on the room side that the DJ use our app and Spotify individually when setting up a queue. This is because the DJ can just use spotify to find a song and add it to the queue without having our App do the whole process (Seeing as our rooms will be virtual rooms and will be in a consistently large number). Users however will be only using our App and won't have to touch Spotify at all (Only to authenticate themselves).
So what's removed is the searching of a song and the ability to go from song to artist to album. Now this code consists of just displaying what song is playing, the artist, album ,cover art and the username of the logged in user along with their display picture.

This should help on preventing too much stress on our App's side when retrieving information.