Pushing as we're currently taking a break and I need to make sure the changes we've made aren't lost. We are currently at the point of making our own buttons that send websockets messages of which button is pressed.
Shao has added:
in \client\src\components\hoc\playerHoc.jsx we need to make functions for each playback option. the functions should simply send a message to the WS server like "LobbyRequest" or "PlayRequest". the WS server in room-service should then handle receiving the request. the WS server should send a message to all the users in the lobby with the info required for each of them to make a Spotify API call (which will probably be just the play/resume api call).