Removed all the main errors. Especially the ones which were constant whenever something was pessed. There apparently is still a "memory leak" , howeve I'm not sure how as it's not affecting anything at all and only happens once.
Also fixed the lobby queue table, before we were missing quite a bit of code fo that and now it's sorted properly.

You'll still notice some warning messages, some are linked with the QR code and "joinwithRoomkey" but it's nothing big, just seems like it needs a dependency, but it doesn't seem to be running poorly without it. Could look into it further. However, apart from that, everything seems nicely bugfree. Especially in mobile view.

--Fixed incorrect names
--Fixed depreciation issues
--Fixed table missing code
--Fixed the mapping where the key was always missing.
--Some other minor fixes too.